@@ -1,27 +1,18 @@
 {
   "name": "worktank-loader",
+  "repository": "github:fabiospampinato/worktank-loader",
   "description": "WebPack plugin for WorkTank which enables you to execute whole files in a worker pool, transparently.",
-  "version": "1.1.0",
+  "version": "1.2.0",
+  "type": "module",
   "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "exports": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
-    "clean": "rimraf dist",
-    "compile": "tsc --skipLibCheck && tstei",
-    "compile:watch": "tsc --skipLibCheck --watch",
+    "clean": "tsex clean",
+    "compile": "tsex compile",
+    "compile:watch": "tsex compile --watch",
     "test": "cd test && webpack && node dist/main.js",
     "prepublishOnly": "npm run clean && npm run compile && npm run test"
-  },
-  "bugs": {
-    "url": "https://github.com/fabiospampinato/worktank-loader/issues"
-  },
-  "license": "MIT",
-  "author": {
-    "name": "Fabio Spampinato",
-    "email": "spampinabio@gmail.com"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fabiospampinato/worktank-loader.git"
   },
   "keywords": [
     "worktank",
@@ -29,19 +20,20 @@
     "plugin"
   ],
   "dependencies": {
-    "esbuild": "^0.10.2",
-    "find-up": "^5.0.0",
-    "worktank": "^1.2.0"
+    "esbuild": "^0.14.39",
+    "find-up-json": "^2.0.0",
+    "worktank": "^2.0.0"
   },
   "peerDependencies": {
-    "webpack": ">=4.0.0"
+    "webpack": ">=5.0.0"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "^5.2.1",
-    "noop3": "^1000.0.0",
-    "rimraf": "^3.0.0",
-    "typescript": "^3.5.3",
-    "typescript-transform-export-interop": "^1.0.2",
-    "webpack": "^4.43.0"
+    "mime2ext": "^2.0.0",
+    "noop-tag": "^2.0.0",
+    "ts-loader": "^9.3.0",
+    "tsex": "^1.1.2",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2"
   }
 }
