@@ -1,19 +1,19 @@
 
 /* IMPORT */
 
-import {strictEqual} from 'assert';
-import * as os from 'os';
-import {sum, multiply, platform, noop, pool} from './example.worker';
+import {strictEqual} from 'node:assert';
+import os from 'node:os';
+import {sum, multiply, platform, mime2ext, pool} from './example.worker';
 import sumDefault from './example.worker';
 
-/* TEST */
+/* MAIN */
 
 const test = async (): Promise<void> => {
 
   strictEqual ( 10, await sum ( 5, 5 ) );
   strictEqual ( 25, await multiply ( 5, 5 ) );
   strictEqual ( os.platform (), await platform () );
-  strictEqual ( undefined, await noop () );
+  strictEqual ( 'mp3', await mime2ext ( 'audio/mp3' ) );
   strictEqual ( 10, await sumDefault ( 5, 5 ) );
 
   pool.terminate ();
